moved concrete polymorphism builder implementations in own files
